lntest/itest: move harness to new file
This in prep for a bigger move in the next commit.
